requirements.txt
discord.py
timeago
jishaku
psutil
numpy
parsedatetime
aiohttp
asyncpg
googletrans
praw
wand
humanize
dropbox
wavelink
youtube_dl
lyricsgenius
pathlib
asyncio
python-dateutil
PrettyTable
dblpy
statcord.py